+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Switched Annotation to Attributes, implements Test and general code style adjust for Activity REST Endpoint !2975 @BibaltiK
 
 ## Dev/Test/CI stuff
+- Remove obsolete picture upload code !2990 @alex.simm
 
 ## Development documentation